Server/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-xai
langchain-core
langchain-huggingface
huggingface-hub
python-dotenv
numpy
scikit-learn
pymupdf
typing
faiss-cpu
torch
rank-bm25
transformers
beautifulsoup4
httpx
python-dotenv
sentence-transformers
groq
langchain-groq